requirements/local.txt
autopep8
flake8
mypy
pillow
pylint
pytest
python-escpos
